+++ docs/source/pages/changelog.rst
@@ -4,6 +4,20 @@ Changelog
 
 The format is based on `Keep a Changelog <http://keepachangelog.com/en/1.0.0/>`_ and this project adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`_.
 
+
+++++++++++++++++++++++++++
+Version 1.0.1 [2018-02-27]
+++++++++++++++++++++++++++
+
+| **Added**
+| - ``OPTIONS`` handler method for ``BaseEndpoint``
+|
+
+| **Fixed**
+| - Some tests for claims that were not using UTC timestamps
+| - Consistency of docs with ``class_views``
+|
+
 ++++++++++++++++++++++++++
 Version 1.0.0 [2018-02-25]
 ++++++++++++++++++++++++++